Some changes
Returned request vouch on vouches page and user reports on user reports page.
