+++ nextjs-app/app/api/common/clientUtils.ts
@@ -1,0 +1,12 @@
+import { ApiCallerConfig } from "./interfaces";
+
+export function getApiClientCallerConfig(): ApiCallerConfig {
+  return {
+    origin: "client",
+    withCredentials: true,
+    headers: {
+      "Content-Type": "application/json",
+      Accept: "application/json",
+    },
+  };
+}
