+++ nextjs-app/app/api/common/interfaces.ts
@@ -1,0 +1,6 @@
+import { AxiosRequestConfig } from "axios";
+
+export type Origin = "server" | "client" | undefined;
+export interface ApiCallerConfig extends AxiosRequestConfig<any> {
+  origin?: Origin; // if not specified it will be treated as "server"
+}
