+++ nextjs-app/app/api/register-players/register-players-read/route.ts
