+++ nextjs-app/app/api/report-system/get-reports/caller.ts
@@ -1,11 +1,7 @@
 import axios, { AxiosRequestConfig } from "axios";
 import type { UserReport } from "../../../services/userReport/getUserReports";
-import { baseUrl, getBaseUrl } from "../../../common/constraints";
-
-export type Origin = "server" | "client" | undefined;
-export interface ApiCallerConfig extends AxiosRequestConfig<any> {
-  origin?: Origin; // if not specified it will be treated as "server"
-}
+import { getBaseUrl } from "../../../common/constraints";
+import { ApiCallerConfig } from "../../common/interfaces";
 
 export const apiCallerGetReports = async ({
   config,
